Move graph-related methods to a dedicated module graph.py. Update the create graph command in the cli tool.
